Extracted classes from fault tree holder - (Task #296)
---
Task #296: Refactor data structures in FaultTreeHolder
